feat: show result path & refactor result component
